Fix a minor physics mistake
Fgravity = m * g
Fnet = m * g = m * a

To find an object's acceleration: a = (Fnet/m)

The issue with this code was that I didn't follow the formula for Newtonian gravity.

In the real world, all objects would fall on earth at the same speed, but the old code, heavier object fell slower than lighter objects.