Create Centripetal motion
Fc = (m * v^2)/r

Because Fc = Fg (force of gravity) We can simplify to: g = v^2/r

Because we are solving for velocity, we end up with:
v = sqrt(g*r)

We can plug the pull of gravity and the radius to get the velocity needed for centripetal motion.

Remember the acceleration (in this case gravity) must be applied perpendicular to the velocity (and the other way around), so we set the initial velocity perpendicular to the direction of the planet.