Implement physically based gravitation fields
Although not to scale with the real world (or should i say solar system), this does still use the classical physics equation I learned from my AP Physics class.

F = G(m1 * m2 / r^2)

m1 and m2 are the masses of the objects, G is the gravitational constant of the universe, r is the distance between the two objects, and F is the force applied to both objects.

In this we solve for the force so we do not have to move any variables around algebraically.

Happy Birthday to me by the way.